requirements.txt
streamlit
openai
pymongo
PyMuPDF
fpdf
dotenv
google-generativeai>=0.7.2